% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy.}
\usage{
accuracy(y, u)
}
\arguments{
\item{y}{Numeric vector of actual classes}

\item{x}{Numeric vector of predicted classes}
}
\description{
Number of correct classes / total guesses.
}

