% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_encoding.R
\name{add_encoding}
\alias{add_encoding}
\title{Add an interaction, poly, one-hot style effect that can be created 
with formula objects and `model.matrix`.}
\usage{
add_encoding(formula, data)
}
\arguments{
\item{formula}{formula describing the transformation. 
Of the form: `~ <stuff>`}

\item{data}{data.frame to which encoding will be created and appended}
}
\value{
copy of data with new columns
}
\description{
Add an interaction, poly, one-hot style effect that can be created 
with formula objects and `model.matrix`.
}

