% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{auc}
\alias{auc}
\title{Classic, binomial AUC.}
\usage{
auc(actual, predicted, weight = rep(1, length(actual)))
}
\arguments{
\item{actual}{Vector of binomaial actuals}

\item{predicted}{Vector of numeric predicted in (-Inf, +Inf)}

\item{weight}{Optional vector of numeric weights in (0, +Inf)}
}
\value{
AUC statistic
}
\description{
Classic, binomial AUC.
}

