% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{binomial_deviance}
\alias{binomial_deviance}
\title{Calculate binomial deviance.}
\usage{
binomial_deviance(y, u, w = rep(1, length(y)))
}
\arguments{
\item{y}{numeric vector of actuals}

\item{u}{numeric vector of predictions}

\item{w}{optional weight vector}
}
\description{
Calculate binomial deviance.
}

