% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o.R
\name{h2o_cvdf}
\alias{h2o_cvdf}
\title{Parse the h2o cv grid into something less ridiculous.}
\usage{
h2o_cvdf(grid)
}
\arguments{
\item{grid}{H2OGrid of H2ORegressionModel objects}
}
\value{
tibble of cv performance ordered by lambda
}
\description{
Parse the h2o cv grid into something less ridiculous.
}

