% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o.R
\name{h2o_plotcv}
\alias{h2o_plotcv}
\title{Plot the CV error vis-a-vis plot.cv.glmnet. Except without standard
errors, because h2o doesn't store them?}
\usage{
h2o_plotcv(sane_grid)
}
\arguments{
\item{sane_grid}{Output of h2o_cvdf}
}
\value{
ggplot object
}
\description{
Plot the CV error vis-a-vis plot.cv.glmnet. Except without standard
errors, because h2o doesn't store them?
}

