% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_conform.R
\name{model_conform}
\alias{model_conform}
\title{Form a model matrix with structure forced to correspond to target.}
\usage{
model_conform(object, data, target, ...)
}
\arguments{
\item{object}{formula to be used with `model.matrix`}

\item{data}{data.frame to be used with `model.matrix`}

\item{target}{matrix to which the output matrix structure will be aligned,
or a character vector of colnames that will be forced.}
}
\description{
Form a model matrix with structure forced to correspond to target.
}

