% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{multivariate_auc}
\alias{multivariate_auc}
\title{Multivariate AUC.}
\usage{
multivariate_auc(y, U)
}
\arguments{
\item{y}{Numeric vector of actual classes}

\item{U}{Numeric matrix of response-scale predictions, classes as columns}
}
\value{
AUC statistic
}
\description{
Multivariate AUC.
}

