% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{mv_binomial_deviance}
\alias{mv_binomial_deviance}
\title{Multivariate binomial deviance.}
\usage{
mv_binomial_deviance(y, U, cap = c(0, 1))
}
\arguments{
\item{y}{Numeric vector of actual classes}

\item{U}{Numeric matrix of response-scale predictions, classes as columns}

\item{cap}{Optional two-element vector of the min and max caps for
predictions in U. Useful for neural networks.}
}
\value{
sum of deviances across all classes
}
\description{
Multivariate binomial deviance.
}

