% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_train_predict.R
\name{predict_folds}
\alias{predict_folds}
\title{Get a matrix of predictions for a list of models. Parallelise
with registerDoMC (if you want, I guess).}
\usage{
predict_folds(data, fold_id, models, predict_model)
}
\arguments{
\item{data}{any training data format that can be indexed by [i, ]}

\item{fold_id}{integer sequence denoting folds starting at 1}

\item{predict_model}{a function with signature (model, data) => predictions}
}
\value{
matrix of predictions
}
\description{
Get a matrix of predictions for a list of models. Parallelise
with registerDoMC (if you want, I guess).
}

