% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cross_train_predict.R
\name{train_folds}
\alias{train_folds}
\title{Train crossfold models. Parallelise with registerDoMC.}
\usage{
train_folds(data, fold_id, train_model)
}
\arguments{
\item{data}{any training data format that can be indexed by [i, ]}

\item{fold_id}{integer sequence denoting folds starting at 1}

\item{train_model}{a function with signature data => model}
}
\value{
list of models
}
\description{
Train crossfold models. Parallelise with registerDoMC.
}

