% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb_perf.R
\name{xgb_perf}
\alias{xgb_perf}
\title{Imitate gbm::gbm.perf for an xgboost::xgb.cv object.}
\usage{
xgb_perf(xgb_cv_model)
}
\arguments{
\item{xgb_cv_model}{output data.table of an xgboost::xgb.cv call}
}
\value{
ggplot object
}
\description{
Plots the train and mean CV error by nrounds.
}

