% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/math.R
\name{multivariate_auc}
\alias{multivariate_auc}
\title{Multivariate AUC.}
\usage{
multivariate_auc(y, U, n_classes = nrow(U))
}
\arguments{
\item{y}{numeric vector of y}

\item{U}{matrix of predictions, rows being classes}

\item{n_classes}{optional number of classes. Infers from dim(U).}
}
\value{
AUC statistic
}
\description{
Multivariate AUC.
}

